Moved firebase functions to src/users
